requirements.txt
aiohttp==3.3.2
python-slugify==1.2.5